+++ bindings/kotlin/src/main/kotlin/com/deftio/triepack/TriePack.kt
@@ -10,7 +10,7 @@
 
 package com.deftio.triepack
 
-const val VERSION = "1.3.1"
+const val VERSION = "1.3.2"
 
 /**
  * Version of the on-disk .trp format this implementation writes. Distinct